RavenDB-20262 /databases/*/indexes endpoint does not require nodeTag now
