+++ src/Raven.Server/Documents/Handlers/Processors/Indexes/IndexHandlerProcessorForGetAll.cs
@@ -1,13 +1,10 @@
 ﻿using System.Linq;
-using System.Net;
 using System.Threading.Tasks;
 using JetBrains.Annotations;
 using Raven.Client.Documents.Indexes;
-using Raven.Server.Json;
 using Raven.Server.ServerWide;
 using Raven.Server.ServerWide.Context;
 using Raven.Server.Web.Http;
-using Sparrow.Json;
 
 namespace Raven.Server.Documents.Handlers.Processors.Indexes;
 
@@ -17,18 +14,7 @@ internal class IndexHandlerProcessorForGetAll : AbstractIndexHandlerProcessorFor
     {
     }
 
-    protected override bool SupportsCurrentNode => true;
-
-    protected override ValueTask HandleCurrentNodeAsync()
-    {
-        var name = GetName();
-        var start = RequestHandler.GetStart();
-        var pageSize = RequestHandler.GetPageSize();
-
-        var indexDefinitions = GetIndexDefinitions(RequestHandler, name, start, pageSize);
-
-        return WriteResultAsync(indexDefinitions);
-    }
+    protected override IndexDefinition[] GetIndexDefinitions(string indexName, int start, int pageSize) => GetIndexDefinitions(RequestHandler, indexName, start, pageSize);
 
     protected override Task HandleRemoteNodeAsync(ProxyCommand<IndexDefinition[]> command, OperationCancelToken token) => RequestHandler.ExecuteRemoteAsync(command, token.Token);
 
@@ -55,25 +41,5 @@ internal class IndexHandlerProcessorForGetAll : AbstractIndexHandlerProcessorFor
         return indexDefinitions;
     }
 
-    private async ValueTask WriteResultAsync(IndexDefinition[] result)
-    {
-        if (result == null)
-        {
-            HttpContext.Response.StatusCode = (int)HttpStatusCode.NotFound;
-            return;
-        }
 
-        using (ContextPool.AllocateOperationContext(out JsonOperationContext context))
-        await using (var writer = new AsyncBlittableJsonTextWriter(context, RequestHandler.ResponseBodyStream()))
-        {
-            writer.WriteStartObject();
-
-            writer.WriteArray(context, "Results", result, (w, c, indexDefinition) =>
-            {
-                w.WriteIndexDefinition(c, indexDefinition);
-            });
-
-            writer.WriteEndObject();
-        }
-    }
 }
